Move frontend assets into ui directory
