+++ frontend/src/ui/assets/fonts/Marianne/Marianne-Bold.woff2
