+++ frontend/src/ui/assets/fonts/Marianne/Marianne-Medium_Italic.woff2
