+++ frontend/src/ui/assets/fonts/Marianne/Marianne-Regular_Italic.woff2
